395 context menu item translate page is sometimes not displayed (#567)
# Problem

We have single context menu for all tabs and dynamically update it when user toggle/refresh tab.

While this update we fetch content script for state of page translation and sometimes it happens too quickly, so content script is not ready to response for requests.

This lead us to an exception and `catch` block is hide context menu in result.

# How this change fixes it

Implemented retries mechanism for query content script.

Now background script can wait response (with reasonable deadline time).

---

* fix(#395): wait response of content script

* chore: update debug print